requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
aioredis >= 2.0.0
aiohttp>=3.7.4
ujson>=4.2.0
pydantic>=1.8.2
pyyaml>=5.4.1
cashews>=2.11.1
starlette>=0.16.0
starlette_prometheus
black~=21.9b0
isort~=5.9.3
httpx~=0.20.0
flake8~=4.0.1
requests~=2.26.0
pytest~=6.2.5
mypy==0.910
mypy-extensions==0.4.3
pytest-asyncio==0.16.0
types-PyYAML==6.0.0
types-ujson==4.2.0
